Retire custom semantic review source
