--- src/supportability_gate/handoff_policy.py
@@ -1,373 +1,0 @@
-"""Cross-check one completion report against authenticated evidence."""
-
-from __future__ import annotations
-
-import re
-import tomllib
-from dataclasses import dataclass
-from typing import Any
-
-_REQUIRED = {
-    "architecture_judgment",
-    "boundary_rationale",
-    "claims",
-    "gate_coverage",
-    "overall_result",
-    "remaining_risks",
-    "responsibility_changes",
-    "simplified_functions",
-    "validation_results",
-}
-_SHA_FIELDS = {"base_sha", "head_sha"}
-_CITATION = re.compile(r"(.+):(\d+)-(\d+)\Z")
-HANDOFF_REPORT_PATH = ".supportability-handoff.toml"
-_REPORT_COMMAND_FIELDS = {"adapter", "arguments", "exit_code"}
-_AUTHORITATIVE_COMMAND_FIELDS = {
-    "adapter",
-    "arguments",
-    "executed",
-    "exit_code",
-}
-_AUTHORITATIVE_OBSERVATION_FIELDS = {
-    "observed_paths",
-    "proof_kind",
-    "zero_statement_paths",
-}
-
-
-class CompletionReportError(ValueError):
-    """Malformed source completion report."""
-
-
-def parse_completion_report(content: bytes) -> dict[str, Any]:
-    """Parse the only supported M10 report document."""
-    try:
-        data = tomllib.loads(content.decode("utf-8"))
-    except (UnicodeDecodeError, tomllib.TOMLDecodeError) as error:
-        raise CompletionReportError("MALFORMED_COMPLETION_REPORT") from error
-    if set(data) != {"schema_version", "completion_report"} or data["schema_version"] != "1.0":
-        raise CompletionReportError("MALFORMED_COMPLETION_REPORT")
-    report = data["completion_report"]
-    if not isinstance(report, dict):
-        raise CompletionReportError("MALFORMED_COMPLETION_REPORT")
-    return report
-
-
-@dataclass(frozen=True)
-class ClaimReview:
-    """Model judgment for one exact report claim."""
-
-    claim_id: str
-    supported: bool
-    citations: tuple[str, ...]
-
-
-def _text_list(value: object) -> bool:
-    return (
-        isinstance(value, list)
-        and bool(value)
-        and all(isinstance(item, str) and bool(item.strip()) for item in value)
-    )
-
-
-def _source_lines(sources: object) -> dict[str, frozenset[int]]:
-    indexed: dict[str, frozenset[int]] = {}
-    if not isinstance(sources, list):
-        return indexed
-    for source in sources:
-        if not isinstance(source, dict) or not isinstance(source.get("path"), str):
-            continue
-        lines = source.get("lines")
-        if not isinstance(lines, list):
-            continue
-        numbers = frozenset(
-            item["line"]
-            for item in lines
-            if isinstance(item, dict) and type(item.get("line")) is int
-        )
-        indexed[source["path"]] = numbers
-    return indexed
-
-
-def completion_citations(report: object) -> tuple[tuple[str, int, int], ...]:
-    """Return well-formed completion claim citations in deterministic order."""
-    if not isinstance(report, dict) or not isinstance(report.get("claims"), list):
-        return ()
-    citations: set[tuple[str, int, int]] = set()
-    for claim in report["claims"]:
-        values = claim.get("citations") if isinstance(claim, dict) else None
-        for citation in values if isinstance(values, list) else []:
-            match = _CITATION.fullmatch(citation) if isinstance(citation, str) else None
-            if match is not None:
-                path, start, end = match.groups()
-                if int(start) <= int(end):
-                    citations.add((path, int(start), int(end)))
-    return tuple(sorted(citations))
-
-
-def _citation_resolves(citation: object, sources: dict[str, frozenset[int]]) -> bool:
-    if not isinstance(citation, str) or (match := _CITATION.fullmatch(citation)) is None:
-        return False
-    path, start_text, end_text = match.groups()
-    start, end = int(start_text), int(end_text)
-    return start <= end and all(line in sources.get(path, ()) for line in range(start, end + 1))
-
-
-def _claims(
-    value: object, sources: dict[str, frozenset[int]], blocks: list[str]
-) -> dict[str, tuple[str, ...]]:
-    claims: dict[str, tuple[str, ...]] = {}
-    if not isinstance(value, list) or not value:
-        blocks.append("MALFORMED_COMPLETION_SECTION:claims")
-        return claims
-    for item in value:
-        if not isinstance(item, dict) or set(item) != {"citations", "id", "text"}:
-            blocks.append("MALFORMED_COMPLETION_SECTION:claims")
-            continue
-        claim_id, text, citations = item["id"], item["text"], item["citations"]
-        if (
-            not isinstance(claim_id, str)
-            or not claim_id.strip()
-            or claim_id in claims
-            or not isinstance(text, str)
-            or not text.strip()
-            or not _text_list(citations)
-        ):
-            blocks.append("MALFORMED_COMPLETION_SECTION:claims")
-            continue
-        trusted = tuple(citations)
-        claims[claim_id] = trusted
-        if not all(_citation_resolves(citation, sources) for citation in trusted):
-            blocks.append(f"UNRESOLVED_COMPLETION_CITATION:{claim_id}")
-    return claims
-
-
-def _commands(
-    value: object,
-    *,
-    authoritative: bool,
-) -> tuple[dict[str, tuple[tuple[str, ...], bool, int]], frozenset[str], bool]:
-    commands: dict[str, tuple[tuple[str, ...], bool, int]] = {}
-    duplicates: set[str] = set()
-    malformed = not isinstance(value, list)
-    if not isinstance(value, list):
-        return commands, frozenset(), malformed
-    expected = _AUTHORITATIVE_COMMAND_FIELDS if authoritative else _REPORT_COMMAND_FIELDS
-    for item in value:
-        fields = set(item) if isinstance(item, dict) else set()
-        if not isinstance(item, dict) or (
-            fields != expected
-            and (not authoritative or fields != expected | _AUTHORITATIVE_OBSERVATION_FIELDS)
-        ):
-            malformed = True
-            continue
-        adapter, arguments, exit_code = (
-            item.get("adapter"),
-            item.get("arguments"),
-            item.get("exit_code"),
-        )
-        if (
-            isinstance(adapter, str)
-            and isinstance(arguments, list)
-            and all(isinstance(argument, str) for argument in arguments)
-            and type(exit_code) is int
-            and (not authoritative or type(item.get("executed")) is bool)
-            and (
-                not authoritative
-                or not fields & _AUTHORITATIVE_OBSERVATION_FIELDS
-                or (
-                    isinstance(item.get("proof_kind"), str)
-                    and isinstance(item.get("observed_paths"), list)
-                    and all(isinstance(path, str) for path in item["observed_paths"])
-                    and isinstance(item.get("zero_statement_paths"), list)
-                    and all(isinstance(path, str) for path in item["zero_statement_paths"])
-                )
-            )
-        ):
-            if adapter in commands:
-                duplicates.add(adapter)
-            commands[adapter] = (tuple(arguments), bool(item.get("executed", True)), exit_code)
-        else:
-            malformed = True
-    return commands, frozenset(duplicates), malformed
-
-
-def _command_blocks(report: dict[str, Any], authoritative: dict[str, Any]) -> list[str]:
-    quality = authoritative.get("quality_profile")
-    observed, observed_duplicates, malformed_observed = _commands(
-        quality.get("commands") if isinstance(quality, dict) else None,
-        authoritative=True,
-    )
-    reported, reported_duplicates, malformed_reported = _commands(
-        report.get("validation_results"), authoritative=False
-    )
-    blocks = [
-        f"INVENTED_VALIDATION_COMMAND:{adapter}" for adapter in reported if adapter not in observed
-    ]
-    if malformed_reported:
-        blocks.append("MALFORMED_VALIDATION_RESULT")
-    if malformed_observed:
-        blocks.append("MALFORMED_AUTHORITATIVE_COMMAND")
-    blocks.extend(f"DUPLICATE_VALIDATION_RESULT:{adapter}" for adapter in reported_duplicates)
-    blocks.extend(f"DUPLICATE_AUTHORITATIVE_COMMAND:{adapter}" for adapter in observed_duplicates)
-    blocks.extend(
-        f"CONTRADICTED_VALIDATION_RESULT:{adapter}"
-        for adapter in reported.keys() & observed.keys()
-        if (reported[adapter][0], reported[adapter][2])
-        != (observed[adapter][0], observed[adapter][2])
-    )
-    blocks.extend(
-        f"AUTHORITATIVE_COMMAND_NOT_EXECUTED:{adapter}"
-        for adapter, (_, executed, _) in observed.items()
-        if not executed
-    )
-    blocks.extend(
-        f"AUTHORITATIVE_COMMAND_FAILED:{adapter}"
-        for adapter, (_, _, exit_code) in observed.items()
-        if exit_code != 0
-    )
-    blocks.extend(
-        f"HIDDEN_FAILED_COMMAND:{adapter}"
-        for adapter, (_, _, exit_code) in observed.items()
-        if exit_code != 0 and adapter not in reported
-    )
-    blocks.extend(
-        f"MISSING_VALIDATION_RESULT:{adapter}" for adapter in observed if adapter not in reported
-    )
-    return blocks
-
-
-def _architecture_result(authoritative: dict[str, Any]) -> str:
-    architecture = authoritative.get("architecture")
-    if not isinstance(architecture, dict) or not architecture.get("executed"):
-        return "MISSING"
-    return "BLOCK" if architecture.get("blocks") else "PASS"
-
-
-def _function_names(authoritative: dict[str, Any]) -> list[str]:
-    functions = authoritative.get("functions")
-    if not isinstance(functions, list):
-        return []
-    return sorted(
-        item["head"]["qualified_name"]
-        for item in functions
-        if isinstance(item, dict)
-        and isinstance(item.get("head"), dict)
-        and isinstance(item["head"].get("qualified_name"), str)
-    )
-
-
-def _risk_blocks(value: object) -> list[str]:
-    if not isinstance(value, list) or not _text_list(value):
-        return ["MALFORMED_COMPLETION_SECTION:remaining_risks"]
-    normalized = {str(item).strip().lower().rstrip(".") for item in value}
-    false_none = {"none", "no risk", "no risks", "no known risk", "no remaining risk"}
-    false_pattern = re.compile(r"\b(?:no|none|zero)\b.*\brisks?\b")
-    return (
-        ["FALSE_NO_REMAINING_RISK"]
-        if normalized & false_none or any(false_pattern.search(item) for item in normalized)
-        else []
-    )
-
-
-def _result_blocks(report: dict[str, Any], authoritative: dict[str, Any]) -> list[str]:
-    blocks: list[str] = []
-    sha_field = next(iter(set(report) & _SHA_FIELDS))
-    if report[sha_field] != authoritative.get(sha_field):
-        blocks.append("STALE_COMPLETION_REPORT_SHA")
-    if authoritative.get("overall_result") != "PASS":
-        blocks.append("AUTHORITATIVE_RESULT_NOT_PASS")
-    if report["overall_result"] != authoritative.get("overall_result"):
-        blocks.append("CONTRADICTED_COMPLETION_RESULT")
-    if report["architecture_judgment"] != _architecture_result(authoritative):
-        blocks.append("CONTRADICTED_ARCHITECTURE_JUDGMENT")
-    if report["gate_coverage"] != authoritative.get("gate_coverage"):
-        blocks.append("CONTRADICTED_GATE_COVERAGE")
-    simplified = report["simplified_functions"]
-    if (
-        isinstance(simplified, list)
-        and all(isinstance(item, str) and bool(item.strip()) for item in simplified)
-        and sorted(simplified) != _function_names(authoritative)
-    ):
-        blocks.append("CONTRADICTED_SIMPLIFIED_FUNCTIONS")
-    return blocks
-
-
-def _review_blocks(
-    claims: dict[str, tuple[str, ...]], reviews: tuple[ClaimReview, ...]
-) -> list[str]:
-    indexed = {review.claim_id: review for review in reviews}
-    blocks = [
-        f"MISSING_COMPLETION_CLAIM_REVIEW:{claim_id}"
-        for claim_id in claims
-        if claim_id not in indexed
-    ]
-    blocks.extend(
-        f"UNBOUND_COMPLETION_CLAIM_REVIEW:{claim_id}"
-        for claim_id in indexed
-        if claim_id not in claims or indexed[claim_id].citations != claims.get(claim_id)
-    )
-    blocks.extend(
-        f"UNSUPPORTED_COMPLETION_CLAIM:{claim_id}"
-        for claim_id, review in indexed.items()
-        if claim_id in claims and not review.supported
-    )
-    return blocks
-
-
-def deterministic_completion_blocks(
-    report: object,
-    authoritative: object,
-    completion_sources: object,
-) -> tuple[str, ...]:
-    """Return machine-resolvable M10 completion-report blocks."""
-    if not isinstance(report, dict) or not isinstance(authoritative, dict):
-        return ("MALFORMED_COMPLETION_REPORT",)
-    missing = sorted(_REQUIRED - set(report))
-    if missing:
-        return (f"MISSING_COMPLETION_SECTION:{missing[0]}",)
-    sha_fields = set(report) & _SHA_FIELDS
-    if not sha_fields:
-        return ("MISSING_COMPLETION_SECTION:base_sha",)
-    if len(sha_fields) != 1:
-        return ("MALFORMED_COMPLETION_SECTION:sha",)
-    unknown = sorted(set(report) - _REQUIRED - sha_fields)
-    if unknown:
-        return (f"MALFORMED_COMPLETION_SECTION:{unknown[0]}",)
-    blocks: list[str] = []
-    for section in ("boundary_rationale", "responsibility_changes"):
-        if not _text_list(report[section]):
-            blocks.append(f"MALFORMED_COMPLETION_SECTION:{section}")
-    simplified = report["simplified_functions"]
-    if not isinstance(simplified, list) or any(
-        not isinstance(item, str) or not item.strip() for item in simplified
-    ):
-        blocks.append("MALFORMED_COMPLETION_SECTION:simplified_functions")
-    blocks.extend(_result_blocks(report, authoritative))
-    blocks.extend(_risk_blocks(report["remaining_risks"]))
-    blocks.extend(_command_blocks(report, authoritative))
-    _claims(report["claims"], _source_lines(completion_sources), blocks)
-    return tuple(sorted(set(blocks)))
-
-
-def evaluate_completion_report(
-    report: object,
-    authoritative: object,
-    completion_sources: object,
-    claim_reviews: tuple[ClaimReview, ...],
-) -> tuple[str, ...]:
-    """Return deterministic and model-backed M10 report blocks."""
-    blocks = list(deterministic_completion_blocks(report, authoritative, completion_sources))
-    if not isinstance(report, dict):
-        return tuple(blocks)
-    claims: list[str] = []
-    indexed: dict[str, tuple[str, ...]] = {}
-    _claims(report.get("claims"), _source_lines(completion_sources), claims)
-    raw_claims = report.get("claims")
-    for item in raw_claims if isinstance(raw_claims, list) else []:
-        if isinstance(item, dict) and isinstance(item.get("id"), str):
-            citations = item.get("citations")
-            if isinstance(citations, list) and all(isinstance(value, str) for value in citations):
-                indexed[item["id"]] = tuple(citations)
-    blocks.extend(_review_blocks(indexed, claim_reviews))
-    return tuple(sorted(set(blocks)))
