--- tests/characterization/m10-review-handoff.golden.json
@@ -1,6 +1,0 @@
-{
-  "clean": [],
-  "unsupported": [
-    "UNSUPPORTED_COMPLETION_CLAIM:claim-1"
-  ]
-}
