--- tests/characterization/m9-reviewer-quality-boundary.golden.json
@@ -1,4 +1,0 @@
-{
-  "quality_runner_contract": true,
-  "semantic_verdict": "PASS"
-}
