+++ tests/characterization/quality-runner-boundary.golden.json
@@ -1,0 +1,3 @@
+{
+  "quality_runner_contract": true
+}
