+++ tests/characterization/refactor-policy-retained.golden.json
