--- tests/characterization/s01-review-state.golden.json
@@ -1,7 +1,0 @@
-{
-  "blocks": [
-    "UNRESOLVED_REVIEW_THREAD:thread-1"
-  ],
-  "body_sha256": "527ea9cd91e27e81543aa5f31c76c4264b09c0f0b4f5531ece166a3bae9b6629",
-  "deterministic": true
-}
