--- tests/characterization/s02-review-events.golden.json
@@ -1,7 +1,0 @@
-{
-  "authenticated_identity": [
-    "mbh-solutions/supportability-gate",
-    67
-  ],
-  "forged_signature": "WEBHOOK_AUTHENTICATION_FAILURE"
-}
