--- tests/characterization/semantic-diagnostics.golden.json
@@ -1,4 +1,0 @@
-{
-  "response_sha256": "54ee03b51653219f6d2e2f372043364360ae78bd2f2b2e920a4d80eee8b04d24",
-  "status": "completed"
-}
