--- tests/characterization/semantic-transport.golden.json
@@ -1,8 +1,0 @@
-{
-  "authorization": true,
-  "method": "POST",
-  "model": "gpt-5.6-sol",
-  "response_status": "completed",
-  "timeout_seconds": 1.0,
-  "url": "http://127.0.0.1:8317/v1/responses"
-}
